# apps/api/src/swagger/integration-api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/integration/tenants": {
      "get": {
        "operationId": "IntegrationController_getTenants",
        "summary": "List all tenants for integration selection",
        "description": "Returns all tenants accessible to the current integration API key.\n\nUse this endpoint first to discover which tenant IDs you can operate on. For the rest of integration endpoints, pass one of those IDs in the X-Tenant-Id header.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "Integration API key used to authenticate every integration request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Tenant ID is optional for this endpoint.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTenantsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid integration API key."
          },
          "403": {
            "description": "Authenticated integration key owner is not authorized."
          }
        },
        "tags": [
          "Integration"
        ]
      }
    },
    "/api/integration/users": {
      "get": {
        "operationId": "IntegrationController_getUsers",
        "summary": "List users for integration",
        "description": "Returns users from the tenant selected by X-Tenant-Id.\n\nSupports keyword search, role filtering, archived filtering ('true' or 'false'), group filtering, sorting, and pagination so integrations can sync user directories in batches.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "Integration API key used to authenticate every integration request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleSlug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "firstName",
                  "type": "string"
                },
                {
                  "const": "lastName",
                  "type": "string"
                },
                {
                  "const": "email",
                  "type": "string"
                },
                {
                  "const": "createdAt",
                  "type": "string"
                },
                {
                  "const": "groupName",
                  "type": "string"
                },
                {
                  "const": "-firstName",
                  "type": "string"
                },
                {
                  "const": "-lastName",
                  "type": "string"
                },
                {
                  "const": "-email",
                  "type": "string"
                },
                {
                  "const": "-createdAt",
                  "type": "string"
                },
                {
                  "const": "-groupName",
                  "type": "string"
                }
              ]
            }
          },
          {
            "name": "groups",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Target tenant context for this request. Must be one of the tenant IDs returned by GET /integration/tenants.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid integration API key."
          },
          "403": {
            "description": "Authenticated integration key owner is not authorized."
          }
        },
        "tags": [
          "Integration"
        ]
      },
      "post": {
        "operationId": "IntegrationController_createUser",
        "summary": "Create user via integration API",
        "description": "Creates a new user inside the tenant selected by X-Tenant-Id.\n\nProvide the user payload in the request body. The response returns the created user ID so your external system can persist the Mentingo mapping.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "Integration API key used to authenticate every integration request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Target tenant context for this request. Must be one of the tenant IDs returned by GET /integration/tenants.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid integration API key."
          },
          "403": {
            "description": "Authenticated integration key owner is not authorized."
          }
        },
        "tags": [
          "Integration"
        ]
      }
    },
    "/api/integration/training-results": {
      "get": {
        "operationId": "IntegrationController_getTrainingResults",
        "summary": "Get training results for integration reporting",
        "description": "Returns training results in JSON for the tenant selected by X-Tenant-Id.\n\nEach response row represents one student-course pair.\n\nUse scope=tenant to list all rows for the tenant, scope=student to list rows for a specific student (studentId required), and scope=course to list rows for a specific course (courseId required). Optional extra filter can further narrow results.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "Integration API key used to authenticate every integration request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": true,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "tenant",
                  "type": "string"
                },
                {
                  "const": "student",
                  "type": "string"
                },
                {
                  "const": "course",
                  "type": "string"
                }
              ]
            }
          },
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Target tenant context for this request. Must be one of the tenant IDs returned by GET /integration/tenants.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrainingResultsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid integration API key."
          },
          "403": {
            "description": "Authenticated integration key owner is not authorized."
          }
        },
        "tags": [
          "Integration"
        ]
      }
    },
...
    "/api/integration/groups": {
      "get": {
        "operationId": "IntegrationController_getGroups",
        "summary": "List groups for integration",
        "description": "Returns groups from the tenant selected by X-Tenant-Id.\n\nSupports keyword filtering, sorting, and pagination so your integration can resolve external groups to Mentingo group IDs before assignment or enrollment operations.",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "Integration API key used to authenticate every integration request.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "name",
                  "type": "string"
                },
                {
                  "const": "createdAt",
                  "type": "string"
                }
              ]
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "const": "en",
                  "type": "string"
                },
                {
                  "const": "pl",
                  "type": "string"
                },
                {
                  "const": "de",
                  "type": "string"
                },
                {
                  "const": "lt",
                  "type": "string"
                },
                {
                  "const": "cs",
                  "type": "string"
                }
              ]
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Target tenant context for this request. Must be one of the tenant IDs returned by GET /integration/tenants.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGroupsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid integration API key."
          },
          "403": {
            "description": "Authenticated integration key owner is not authorized."
          }
        },
        "tags": [
          "Integration"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mentingo API",
    "description": "This is the API documentation for Mentingo",
    "version": "v0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "GetTenantsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "format": "uuid",
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "host": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "host"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
      "CreateUserBody": {
        "type": "object",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "roleSlugs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "anyOf": [
              {
                "const": "en",
                "type": "string"
              },
              {
                "const": "pl",
                "type": "string"
              },
              {
                "const": "de",
                "type": "string"
              },
              {
                "const": "lt",
                "type": "string"
              },
              {
                "const": "cs",
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "roleSlugs"
        ]
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}